# Perf sur query des multivalidations