Je réécris la query qui permet d'aller chercher les multivalidations et métadonnées associées à un dataset. Au lieu d'utiliser des join, j'utilise des lateral joins, et le résultat est...comment dire, meilleur.

Ca devrait bien accélerer le chargement des pages datasets qui sont lentes actuellement.

Avant : [debug] QUERY OK source="resource" **db=911.4ms** idle=1067.8ms
SELECT r0."id", s2."id", s2."validation_timestamp", s2."validator", s2."validator_version", s2."command", s2."result", s2."data_vis", s2."max_error", s2."oban_args", s2."resource_id", s2."resource_history_id", s2."validated_data_name", s2."secondary_resource_id", s2."secondary_resource_history_id", s2."secondary_validated_data_name", s2."inserted_at", s2."updated_at", r3."id", r3."resource_id", r3."resource_history_id", r3."multi_validation_id", r3."metadata", r3."modes", r3."features", r3."inserted_at", r3."updated_at" FROM "resource" AS r0 LEFT OUTER JOIN (SELECT DISTINCT ON (sr0."resource_id") sr0."id" AS "id", sr0."datagouv_id" AS "datagouv_id", sr0."payload" AS "payload", sr0."last_up_to_date_at" AS "last_up_to_date_at", sr0."inserted_at" AS "inserted_at", sr0."updated_at" AS "updated_at", sr0."resource_id" AS "resource_id" FROM "resource_history" AS sr0 ORDER BY sr0."resource_id", sr0."inserted_at" DESC) AS s1 ON s1."resource_id" = r0."id" LEFT OUTER JOIN (SELECT DISTINCT ON (sm0."resource_history_id", sm0."resource_id", sm0."validator") sm0."id" AS "id", sm0."validation_timestamp" AS "validation_timestamp", sm0."validator" AS "validator", sm0."validator_version" AS "validator_version", sm0."command" AS "command", sm0."result" AS "result", sm0."data_vis" AS "data_vis", sm0."max_error" AS "max_error", sm0."oban_args" AS "oban_args", sm0."resource_id" AS "resource_id", sm0."resource_history_id" AS "resource_history_id", sm0."validated_data_name" AS "validated_data_name", sm0."secondary_resource_id" AS "secondary_resource_id", sm0."secondary_resource_history_id" AS "secondary_resource_history_id", sm0."secondary_validated_data_name" AS "secondary_validated_data_name", sm0."inserted_at" AS "inserted_at", sm0."updated_at" AS "updated_at" FROM "multi_validation" AS sm0 WHERE (sm0."validator" = ANY($1)) ORDER BY sm0."resource_history_id", sm0."resource_id", sm0."validator", sm0."resource_history_id" DESC, sm0."resource_id" DESC, sm0."validator" DESC, sm0."validation_timestamp" DESC) AS s2 ON (s1."id" = s2."resource_history_id") OR (r0."id" = s2."resource_id") LEFT OUTER JOIN "resource_metadata" AS r3 ON r3."multi_validation_id" = s2."id" WHERE (r0."dataset_id" = $2) [["GTFS transport-validator", "gtfs-realtime-validator", "validata-api", "EXJSONSchema", "MobilityData/gbfs-validator"], 893]

Après : [debug] QUERY OK source="resource" **db=1.9ms** idle=431.4ms
SELECT r0."id", s2."id", s2."validation_timestamp", s2."validator", s2."validator_version", s2."command", s2."result", s2."data_vis", s2."max_error", s2."oban_args", s2."resource_id", s2."resource_history_id", s2."validated_data_name", s2."secondary_resource_id", s2."secondary_resource_history_id", s2."secondary_validated_data_name", s2."inserted_at", s2."updated_at", s3."id", s3."resource_id", s3."resource_history_id", s3."multi_validation_id", s3."metadata", s3."modes", s3."features", s3."inserted_at", s3."updated_at" FROM "resource" AS r0 LEFT OUTER JOIN LATERAL (SELECT sr0."id" AS "id", sr0."datagouv_id" AS "datagouv_id", sr0."payload" AS "payload", sr0."last_up_to_date_at" AS "last_up_to_date_at", sr0."inserted_at" AS "inserted_at", sr0."updated_at" AS "updated_at", sr0."resource_id" AS "resource_id" FROM "resource_history" AS sr0 WHERE (r0."id" = sr0."resource_id") ORDER BY sr0."inserted_at" DESC LIMIT 1) AS s1 ON TRUE LEFT OUTER JOIN LATERAL (SELECT DISTINCT ON (sm0."resource_history_id", sm0."resource_id", sm0."validator") sm0."id" AS "id", sm0."validation_timestamp" AS "validation_timestamp", sm0."validator" AS "validator", sm0."validator_version" AS "validator_version", sm0."command" AS "command", sm0."result" AS "result", sm0."data_vis" AS "data_vis", sm0."max_error" AS "max_error", sm0."oban_args" AS "oban_args", sm0."resource_id" AS "resource_id", sm0."resource_history_id" AS "resource_history_id", sm0."validated_data_name" AS "validated_data_name", sm0."secondary_resource_id" AS "secondary_resource_id", sm0."secondary_resource_history_id" AS "secondary_resource_history_id", sm0."secondary_validated_data_name" AS "secondary_validated_data_name", sm0."inserted_at" AS "inserted_at", sm0."updated_at" AS "updated_at" FROM "multi_validation" AS sm0 WHERE ((r0."id" = sm0."resource_id") OR (s1."id" = sm0."resource_history_id")) AND (sm0."validator" = ANY($1)) ORDER BY sm0."resource_history_id", sm0."resource_id", sm0."validator", sm0."validation_timestamp" DESC) AS s2 ON TRUE LEFT OUTER JOIN LATERAL (SELECT sr0."id" AS "id", sr0."resource_id" AS "resource_id", sr0."resource_history_id" AS "resource_history_id", sr0."multi_validation_id" AS "multi_validation_id", sr0."metadata" AS "metadata", sr0."modes" AS "modes", sr0."features" AS "features", sr0."inserted_at" AS "inserted_at", sr0."updated_at" AS "updated_at" FROM "resource_metadata" AS sr0 WHERE (s2."id" = sr0."multi_validation_id")) AS s3 ON TRUE WHERE (r0."dataset_id" = $2) [["GTFS transport-validator", "gtfs-realtime-validator", "validata-api", "EXJSONSchema", "MobilityData/gbfs-validator"], 893]